criado Exercicio2

+++ Lista1/src/com/dbc/Exercicio2.java
@@ -6,15 +6,15 @@ public class Exercicio2 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Digite a PRIMEIRA nota: ");
-        float nota1 = scanner.nextInt();
+        double nota1 = scanner.nextDouble();
         System.out.println("Digite a SEGUNDA nota: ");
-        float nota2 = scanner.nextInt();
+        double nota2 = scanner.nextDouble();
         System.out.println("Digite a TERCEIRA nota: ");
-        float nota3 = scanner.nextInt();
+        double nota3 = scanner.nextDouble();
         System.out.println("Digite a QUARTA nota: ");
-        float nota4 = scanner.nextInt();
+        double nota4 = scanner.nextDouble();
 
-        float media = (nota1 + nota2 + nota3 + nota4) / 4;
+        double media = (nota1 + nota2 + nota3 + nota4) / 4;
 
         if (media <=10 && media >= 7) {
             System.out.println("Sua média foi " + media + ". Você foi APROVADO!");
